+++ src/Resources/app/administration/src/module/ui-feed-export/component/ui-feed-export-form/index.js
@@ -5,12 +5,16 @@ const { Component, Mixin } = Shopware;
 Component.register('ui-feed-export-form', {
     template,
 
+    inject: ['systemConfigApiService'],
+
     data() {
         return {
             salesChannelValue: null,
             salesChannelLanguageValue: null,
             exportTypeValue: null,
             typeSelectOptions: [],
+            isCacheDisable: false,
+            isLoadingCache: false,
         }
     },
 
@@ -18,6 +22,7 @@ Component.register('ui-feed-export-form', {
         Mixin.getByName('notification')
     ],
     mounted () {
+        this.getPluginConfig()
         this.getExportTypeValues()
     },
     filters: {
@@ -28,6 +33,10 @@ Component.register('ui-feed-export-form', {
         }
     },
     methods: {
+         async getPluginConfig() {
+             const config = await this.systemConfigApiService.getValues('OmikronFactFinder.config');
+             this.isCacheDisable = config['OmikronFactFinder.config.enableExportCache'];
+        },
         getExportTypeValues() {
             const httpClient = Shopware.Service('syncService').httpClient;
             const url = '_action/fact-finder/get-export-type-options';
@@ -56,6 +65,16 @@ Component.register('ui-feed-export-form', {
                 message: this.$tc('ui-feed-export.component.export_form.alert_error.text')
             })
         },
+        successRefreshCacheWindow() {
+            this.createNotificationSuccess({
+                message: this.$tc('ui-feed-export.component.export_form.refresh_cache_success.text')
+            })
+        },
+        errorRefreshCacheWindow() {
+            this.createNotificationError({
+                message: this.$tc('ui-feed-export.component.export_form.refresh_cache_error.text')
+            })
+        },
         getFeedExportFile(url) {
             const httpClient = Shopware.Service('syncService').httpClient;
             const basicHeaders = {
@@ -79,6 +98,33 @@ Component.register('ui-feed-export-form', {
                     } else {
                         this.errorFeedGenerationWindow();
                     }
+                });
+        },
+        refreshExportCache(url) {
+            this.isLoadingCache = true;
+            const httpClient = Shopware.Service('syncService').httpClient;
+            const basicHeaders = {
+                Authorization: `Bearer ${Shopware.Context.api.authToken.access}`,
+                'Content-Type': 'application/json'
+            };
+            const params = {
+                salesChannelValue: this.salesChannelValue,
+                salesChannelLanguageValue: this.salesChannelLanguageValue
+            };
+
+            httpClient
+                .get(url, {
+                    headers: basicHeaders,
+                    params: params
+                })
+                .then((response) => {
+                    if (response.status === 200) {
+                        this.successRefreshCacheWindow();
+                    } else {
+                        this.errorRefreshCacheWindow();
+                    }
+
+                    this.isLoadingCache = false;
                 });
         }
     }